Reformat project: create struct of directs: cmd,pkg.internal, add .gitignore
