+++ cmd/app/main.go
@@ -1,0 +1,91 @@
+package main
+
+import (
+	"database/sql"
+	"fmt"
+
+	_ "modernc.org/sqlite"
+
+	"github.com/Ekvo/yandex-practicum-golang-ci-cd-final/internal/services"
+	"github.com/Ekvo/yandex-practicum-golang-ci-cd-final/internal/source"
+)
+
+func main() {
+	db, err := sql.Open("sqlite", "./storage/tracker.db")
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	defer db.Close()
+
+	store := source.NewParcelStore(db)
+	service := services.NewParcelService(store)
+
+	// регистрация посылки
+	client := 1
+	address := "Псков, д. Пушкина, ул. Колотушкина, д. 5"
+	p, err := service.Register(client, address)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// изменение адреса
+	newAddress := "Саратов, д. Верхние Зори, ул. Козлова, д. 25"
+	err = service.ChangeAddress(p.Number, newAddress)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// изменение статуса
+	err = service.NextStatus(p.Number)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// вывод посылок клиента
+	err = service.PrintClientParcels(client)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// попытка удаления отправленной посылки
+	err = service.Delete(p.Number)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// вывод посылок клиента
+	// предыдущая посылка не должна удалиться, так как её статус «НЕ «зарегистрирована»
+	err = service.PrintClientParcels(client)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// регистрация новой посылки
+	p, err = service.Register(client, address)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// удаление новой посылки
+	err = service.Delete(p.Number)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+
+	// вывод посылок клиента
+	// здесь не должно быть последней посылки, так как она должна была успешно удалиться
+	err = service.PrintClientParcels(client)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+}
