+++ internal/model/models.go
@@ -1,0 +1,9 @@
+package model
+
+type Parcel struct {
+	Number    int
+	Client    int
+	Status    string
+	Address   string
+	CreatedAt string
+}
