+++ internal/services/logicservices.go
@@ -1,0 +1,75 @@
+package services
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/Ekvo/yandex-practicum-golang-ci-cd-final/internal/model"
+	vr "github.com/Ekvo/yandex-practicum-golang-ci-cd-final/internal/variables"
+)
+
+func (s ParcelService) Register(client int, address string) (model.Parcel, error) {
+	parcel := model.Parcel{
+		Client:    client,
+		Status:    vr.ParcelStatusRegistered,
+		Address:   address,
+		CreatedAt: time.Now().UTC().Format(time.RFC3339),
+	}
+
+	id, err := s.store.Add(parcel)
+	if err != nil {
+		return parcel, err
+	}
+
+	parcel.Number = id
+
+	fmt.Printf("Новая посылка № %d на адрес %s от клиента с идентификатором %d зарегистрирована %s\n",
+		parcel.Number, parcel.Address, parcel.Client, parcel.CreatedAt)
+
+	return parcel, nil
+}
+
+func (s ParcelService) PrintClientParcels(client int) error {
+	parcels, err := s.store.GetByClient(client)
+	if err != nil {
+		return err
+	}
+
+	fmt.Printf("Посылки клиента %d:\n", client)
+	for _, parcel := range parcels {
+		fmt.Printf("Посылка № %d на адрес %s от клиента с идентификатором %d зарегистрирована %s, статус %s\n",
+			parcel.Number, parcel.Address, parcel.Client, parcel.CreatedAt, parcel.Status)
+	}
+	fmt.Println()
+
+	return nil
+}
+
+func (s ParcelService) NextStatus(number int) error {
+	parcel, err := s.store.Get(number)
+	if err != nil {
+		return err
+	}
+
+	var nextStatus string
+	switch parcel.Status {
+	case vr.ParcelStatusRegistered:
+		nextStatus = vr.ParcelStatusSent
+	case vr.ParcelStatusSent:
+		nextStatus = vr.ParcelStatusDelivered
+	case vr.ParcelStatusDelivered:
+		return nil
+	}
+
+	fmt.Printf("У посылки № %d новый статус: %s\n", number, nextStatus)
+
+	return s.store.SetStatus(number, nextStatus)
+}
+
+func (s ParcelService) ChangeAddress(number int, address string) error {
+	return s.store.SetAddress(number, address)
+}
+
+func (s ParcelService) Delete(number int) error {
+	return s.store.Delete(number)
+}
