+++ internal/source/source.go
@@ -1,0 +1,11 @@
+package source
+
+import "database/sql"
+
+type ParcelStore struct {
+	db *sql.DB
+}
+
+func NewParcelStore(db *sql.DB) ParcelStore {
+	return ParcelStore{db: db}
+}
