+++ internal/variables/variables.go
@@ -1,0 +1,7 @@
+package variables
+
+const (
+	ParcelStatusRegistered = "registered"
+	ParcelStatusSent       = "sent"
+	ParcelStatusDelivered  = "delivered"
+)
